requirements.txt
# Core dependencies
requests>=2.31.0
websocket-client>=1.6.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Financial calculations
scipy>=1.11.0
py_vollib>=1.0.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data validation
pydantic>=2.0.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Web framework (for dashboard)
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Plotting and visualization
plotly>=5.15.0
dash>=2.12.0

# Configuration management
pyyaml>=6.0
